Move RedisLock to a common location.
- Fixes a Python 3 issue with the implementation that prevented
  releasing the lock manually.

- Increase the lock validity to 30 minutes for safety:

    If the lock expires while a task is still legitimately running,
    another competing task could be started, which could result in
    database inconsistencies.
