+++ server/ec2spotmanager/tests/test_task_graph.py
@@ -30,7 +30,7 @@ def test_update_pool_graph(mocker):
     mock_group = mocker.patch('celery.group')
     mock_chain = mocker.patch('celery.chain')
     mock_chord = mocker.patch('celery.chord')
-    mock_lock = mocker.patch('ec2spotmanager.cron._RedisLock')
+    mock_lock = mocker.patch('ec2spotmanager.cron.RedisLock')
     mock_check_and_resize_pool = mocker.patch('ec2spotmanager.tasks.check_and_resize_pool')
     mock_cycle_and_terminate_disabled = mocker.patch('ec2spotmanager.tasks.cycle_and_terminate_disabled')
     mock_terminate_instances = mocker.patch('ec2spotmanager.tasks.terminate_instances')
@@ -154,7 +154,7 @@ def test_update_pool_graph_unsupported_running(mocker):
     mock_group = mocker.patch('celery.group')
     mock_chain = mocker.patch('celery.chain')
     mock_chord = mocker.patch('celery.chord')
-    mock_lock = mocker.patch('ec2spotmanager.cron._RedisLock')
+    mock_lock = mocker.patch('ec2spotmanager.cron.RedisLock')
     mock_check_and_resize_pool = mocker.patch('ec2spotmanager.tasks.check_and_resize_pool')
     mock_cycle_and_terminate_disabled = mocker.patch('ec2spotmanager.tasks.cycle_and_terminate_disabled')
     mock_terminate_instances = mocker.patch('ec2spotmanager.tasks.terminate_instances')
